# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FlightMovement API Documentation",
        "description": "API documentation for FlightMovement",
        "contact": {
            "email": "dev@camranh.aero"
        },
        "version": "1.0.0"
    },
    "servers": [
        {

            "url": "http://localhost:8000/api",
            "description": "Rest API"
        }
    ],
    "paths": {
...
        "/process-barcode": {
            "get": {
                "tags": [
                    "Barcode Processing"
                ],
                "summary": "Process barcode data for baggage recheck",
                "description": "Processes barcode scan data to determine if baggage recheck is required. First checks TagRecheck table, then DepartureMovementView if not found.",
                "operationId": "9ed12b6c491127c7ce8e9b79402cec04",
                "parameters": [
                    {
                        "name": "airline",
                        "in": "query",
                        "description": "Airline code (e.g., VJ, HY)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "VJ"
                        }
                    },
                    {
                        "name": "flightNumber",
                        "in": "query",
                        "description": "Flight number (e.g., 870, 0562)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "870"
                        }
                    },
                    {
                        "name": "flightDate",
                        "in": "query",
                        "description": "Flight date in Y-m-d format",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-03-29"
                        }
                    },
                    {
                        "name": "seatNumber",
                        "in": "query",
                        "description": "Seat number (e.g., 15F, 009F - leading zeros will be handled)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "15F"
                        }
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "description": "Passenger last name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "KIM"
                        }
                    },
                    {
                        "name": "firstName",
                        "in": "query",
                        "description": "Passenger first name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "SOOJINMS"
                        }
                    },
                    {
                        "name": "languageCode",
                        "in": "query",
                        "description": "Language code for response message (EN, VI, KO, ZH)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "EN",
                                "VI",
                                "KO",
                                "ZH"
                            ],
                            "example": "EN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - Recheck required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "Recheck": {
                                            "type": "string",
                                            "example": "Yes"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "Passenger": {
                                            "type": "string",
                                            "example": "KIM SOOJINMS"
                                        },
                                        "Counter": {
                                            "type": "string",
                                            "example": "H"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Please proceed to the recheck counter at H"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No matching records found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No matching records found for the provided barcode data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing required parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Missing required parameters: airline, flightNumber, flightDate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recheck-statistics/{date}": {
            "get": {
                "tags": [
                    "Barcode Processing"
                ],
                "summary": "Get recheck statistics by date",
                "description": "Returns statistics for baggage recheck operations on a specific date",
                "operationId": "9726a8ceff116ace6e8fd15e705976ae",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "Date in Y-m-d format",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-03-29"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "date": {
                                                    "type": "string",
                                                    "example": "2025-03-29"
                                                },
                                                "total_recheck": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "finished_recheck": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "pending_recheck": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "completion_rate": {
                                                    "type": "number",
                                                    "example": 93.33
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recheck-by-counter/{date}/{counter}": {
            "get": {
                "tags": [
                    "Barcode Processing"
                ],
                "summary": "Get recheck records by counter and date",
                "description": "Returns all recheck records for a specific counter on a specific date",
                "operationId": "44f9543e52a3fa2a72c4d7cc94474df1",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "Date in Y-m-d format",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-03-29"
                        }
                    },
                    {
                        "name": "counter",
                        "in": "path",
                        "description": "Counter letter (e.g., H, C)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "H"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Records retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}